feat: cc-0058 Add delete file
